\name{create_function_df}
\alias{create_function_df}
\title{Create data.frame with information on package functions}
\description{For a given package, return a data.frame with one line per function, containing information from help files for each function}
\usage{create_function_df(pkg_name)}
\arguments{
  \item{pke_name}{Name of an (installed) R package}
}
\value{Returns a \code{data.frame}.}